created two folders in view and moved corresponding files
