+++ client/src/App.jsx
@@ -2,12 +2,12 @@ import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import NavBar from "./components/NavBar/NavBar.jsx";
 import heroImage from "./assets/hero.png";
-import Home from "./views/Home.jsx";
-import Login from "./views/Login.jsx";
-import StudentSignUp from "./views/StudentSignUp.jsx";
-import BusinessSignUp from "./views/BusinessSignUp.jsx";
-import Question from "./views/Question.jsx";
-import Task from "./views/Task.jsx";
+import Home from "./views/landingPage/Home.jsx";
+import Login from "./views/landingPage/Login.jsx";
+import StudentSignUp from "./views/landingPage/StudentSignUp.jsx";
+import BusinessSignUp from "./views/landingPage/BusinessSignUp.jsx";
+import Question from "./views/mingleGame//Question.jsx";
+import Task from "./views/mingleGame/Task.jsx";
 
 function App() {
   return (
